+++ README.md
@@ -50,8 +50,6 @@
 - [x] 블랙잭에 필요한 카드 추가
 - [x] 블랙잭에 필요한 덱 추가
 - [x] 덱에서 카드를 뽑는 기능 추가
-- [x] 블랙잭에 필요한 심판 추가
-- [x] 심판이 카드를 제공하는 기능 추가
 - [x] 블랙잭에 참여하는 플레이어 추가
 - [x] 플레이어와 딜러의 손패 추가
 - [x] 플레이어 손패에 카드를 추가하는 기능 추가
@@ -61,3 +59,4 @@
 - [x] 카드의 점수를 계산하는 기능 추가
 - [x] 손패에서 점수를 반환하는 기능 추가
 - [x] 딜러가 카드를 받아서 추가하는 기능 추가
+- [x] 블랙잭 게임을 위해 플레이어와 딜러를 생성하는 기능 추가
